backend/requirements.txt
Django>=5
djangorestframework
psycopg2-binary
celery[redis]
django-celery-beat
django-redis
python-decouple